% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{pop_chart}
\alias{pop_chart}
\title{Plot pie chart of population}
\usage{
pop_chart(dat, county, id_dom = "investChart")
}
\arguments{
\item{dat}{data set}

\item{county}{character commune name}
}
\value{
ndv3 plot discretBarChart
}
\description{
Plot pie chart of population
}

