% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{popup_content}
\alias{popup_content}
\title{Create popup content}
\usage{
popup_content(region, nbr)
}
\arguments{
\item{region}{character}

\item{nbr}{numeric popluation}
}
\value{
character popup content
}
\description{
Create popup content
}

