% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{load_survey}
\alias{load_survey}
\title{Extract survey raw data}
\usage{
load_survey(force = FALSE)
}
\value{
data.frame
}
\description{
Survey data are stored in google spredsheetin Drive.
}
\examples{
load_survey
}

