% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{popup_content}
\alias{popup_content}
\title{Create popup content}
\usage{
popup_content(commune, nbr, depart, arrond)
}
\arguments{
\item{commune}{character}

\item{nbr}{numeric popluation}

\item{depart}{character}

\item{arrond}{character}
}
\value{
character popup content
}
\description{
Create popup content
}

